+++ src/Core/Serialization/XmlSerializer.php
@@ -38,19 +38,11 @@ use DOMElement;
  */
 class XmlSerializer extends BaseSerializer
 {
-    private readonly DOM\NormalizerFactory $normalizerFactory;
-
-    private readonly string $xmlVersion;
-    private readonly string $xmlEncoding;
-
     public function __construct(
-        DOM\NormalizerFactory $normalizerFactory,
-        string $xmlVersion = '1.0',
-        string $xmlEncoding = 'UTF-8'
+        protected readonly DOM\NormalizerFactory $normalizerFactory,
+        protected readonly string $xmlVersion = '1.0',
+        protected readonly string $xmlEncoding = 'UTF-8'
     ) {
-        $this->normalizerFactory = $normalizerFactory;
-        $this->xmlVersion = $xmlVersion;
-        $this->xmlEncoding = $xmlEncoding;
     }
 
     protected function realNormalize(Bom $bom): DOMElement